Extract ResultFragment from ImportFragment
