+++ app/src/main/java/com/etesync/syncadapter/ui/importlocal/LocalContactImportFragment.java
@@ -36,7 +36,7 @@ public class LocalContactImportFragment extends Fragment {
 
     private Account account;
     private CollectionInfo info;
-    private OnImportCallback importCallback;
+    private ResultFragment.OnImportCallback importCallback;
 
     public static LocalContactImportFragment newInstance(Account account, CollectionInfo info) {
         LocalContactImportFragment frag = new LocalContactImportFragment();
@@ -82,7 +82,7 @@ public class LocalContactImportFragment extends Fragment {
         // This makes sure that the container activity has implemented
         // the callback interface. If not, it throws an exception
         try {
-            importCallback = (OnImportCallback) getActivity();
+            importCallback = (ResultFragment.OnImportCallback) getActivity();
         } catch (ClassCastException e) {
             throw new ClassCastException(getActivity().toString()
                     + " must implement MyInterface ");
@@ -96,7 +96,7 @@ public class LocalContactImportFragment extends Fragment {
         // This makes sure that the container activity has implemented
         // the callback interface. If not, it throws an exception
         try {
-            importCallback = (OnImportCallback) activity;
+            importCallback = (ResultFragment.OnImportCallback) activity;
         } catch (ClassCastException e) {
             throw new ClassCastException(activity.toString()
                     + " must implement MyInterface ");
@@ -168,8 +168,7 @@ public class LocalContactImportFragment extends Fragment {
         @Override
         protected void onPostExecute(Boolean result) {
             progressDialog.dismiss();
-            if (result) importCallback.onImportSuccess();
-            else importCallback.onImportFailed();
+            importCallback.onImportResult(new ResultFragment.ImportResult());
         }
 
         private boolean importContacts(LocalAddressBook localAddressBook) {
