--- app/src/main/java/com/etesync/syncadapter/ui/importlocal/OnImportCallback.java
@@ -1,11 +1,0 @@
-package com.etesync.syncadapter.ui.importlocal;
-
-/**
- * Created by tal on 30/03/17.
- */
-
-public interface OnImportCallback {
-    void onImportSuccess();
-
-    void onImportFailed();
-}
